refactor: refactor color logic into hook, update playlist styling
