+++ src/app/layout.tsx
@@ -1,9 +1,6 @@
 import type { Metadata } from 'next';
 import { geistSans, geistMono, corben } from '@/lib/fonts';
 import Header from '@/components/Header/Header';
-import { auth } from '@/lib/auth';
-import { headers } from 'next/headers';
-import { redirect } from 'next/navigation';
 import './globals.css';
 
 export const metadata: Metadata = {
@@ -16,13 +13,6 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const session = await auth.api.getSession({ headers: await headers() });
-
-  // Redirect users if app is in development
-  if (process.env.NODE_ENV === 'development' && session?.user?.id !== process.env.DEV_ID) {
-    redirect('/under-construction');
-  }
-
   return (
     <html lang="en" className="min-w-[280px] md:scroll-smooth">
       <head>
