+++ src/app/playlist/[id]/page.tsx
@@ -23,15 +23,11 @@ export default async function PlaylistPage({ params }: Params) {
   const playlist: PlaylistResponse = await getPlaylist(id);
 
   return (
-    <div className="relative flex h-screen flex-col items-center overflow-hidden p-4">
+    <div className="relative flex h-screen flex-col items-center p-4">
       <DashboardChevron />
-      <div className="flex flex-col items-center gap-6 pt-24 md:pt-16">
-        <h1 className="font-corben text-center text-3xl font-bold md:text-4xl">{playlist.name}</h1>
-        <p className="text-secondary-text text-center">
-          Click on a track to select a different color option.
-        </p>
+      <div className="mt-14 w-full flex-1 overflow-hidden md:mt-18">
+        <SortedPlaylist playlist={playlist} />
       </div>
-      <SortedPlaylist playlist={playlist} />
     </div>
   );
 }
@@ -40,7 +36,7 @@ function DashboardChevron() {
   return (
     <Link
       href="/dashboard"
-      className="hover:bg-white-active active:bg-white-active fixed top-16 left-4 z-5 cursor-pointer rounded-lg p-2 transition-colors duration-300 md:top-20 md:left-20"
+      className="hover:bg-white-active active:bg-white-active fixed top-16 left-1 z-5 cursor-pointer rounded-lg p-2 transition-colors duration-300 md:top-20 md:left-20"
     >
       <svg
         xmlns="http://www.w3.org/2000/svg"
@@ -48,7 +44,7 @@ function DashboardChevron() {
         viewBox="0 0 24 24"
         strokeWidth={2}
         stroke="currentColor"
-        className="text-foreground h-8 w-8"
+        className="text-secondary-text h-8 w-8"
       >
         <path strokeLinecap="round" strokeLinejoin="round" d="M15 19l-7-7 7-7" />
       </svg>
